Fixing types for quantity, limitPrice and stopPrice

+++ src/main/java/it/trade/model/reponse/OrderInfo.java
@@ -10,7 +10,7 @@ public class OrderInfo {
 
         @SerializedName("quantity")
         @Expose
-        public Integer quantity;
+        public Double quantity;
 
         @SerializedName("symbol")
         @Expose
